requirements.txt
asgiref==3.5.2
attrs==22.1.0
Automat==20.2.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
constantly==15.1.0
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1
django-debug-toolbar==3.6.0
gunicorn==20.1.0
heroku==0.1.4
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
oauthlib==3.2.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dateutil==1.5
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sorl-thumbnail==12.8.0
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.11
whitenoise==6.2.0
zope.interface==5.4.0
